Add support for ambiguation after satp is updated
According to the RISC-V manual, writing satp does not imply any
ordering constraints between page-table updates and subsequent
address translations.

Therefore, we track the status of satp write and sfence.vma. If
satp is written and changes from Bare mode to Sv39, we marks it as
an ambiguation state.

Note that the satp_written is too coarse-grained and must be
optimized in the future.